+++ productService/src/main/java/com/myShoppingApp/productService/dto/ProductRequest.java
@@ -1,7 +1,7 @@
-package com.myShoppingApp.dto;
+package com.myShoppingApp.productService.dto;
 
 import java.math.BigDecimal;
 
 public record ProductRequest(String name, String description, BigDecimal price) {
- 
+
 }
