+++ productService/src/main/java/com/myShoppingApp/productService/dto/ProductResponse.java
@@ -1,8 +1,7 @@
-package com.myShoppingApp.dto;
+package com.myShoppingApp.productService.dto;
 
 import java.math.BigDecimal;
 
-
 public record ProductResponse(String id, String name, String description, BigDecimal price) {
-     
+
 }
