--- productService/src/main/java/com/myShoppingApp/repository/IProductRepository.java
@@ -1,9 +1,0 @@
-package com.myShoppingApp.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.myShoppingApp.model.Product;
-
-public interface IProductRepository extends MongoRepository<Product, String>{
-    
-}
